fix(F-496): fix build error
