+++ frontend/src/contexts/Rating.ts
@@ -1,5 +1,18 @@
-import { RatingContextValue } from '@/interfaces/inputs/Rating';
-import React, { useContext } from 'react';
+import React, { useContext, MouseEvent, KeyboardEvent } from 'react';
+
+export type RatingContextValue = {
+  value: number;
+  readOnly: boolean;
+  hoverValue: number | null;
+  focusedStar: number | null;
+  handleValueChange: (
+    event: MouseEvent<HTMLButtonElement> | KeyboardEvent<HTMLButtonElement>,
+    value: number
+  ) => void;
+  handleKeyDown: (event: KeyboardEvent<HTMLButtonElement>) => void;
+  setHoverValue: (value: number | null) => void;
+  setFocusedStar: (value: number | null) => void;
+};
 
 export const RatingContext = React.createContext<RatingContextValue | null>(null);
